feat(web): render all pending ask questions, not just the active session
With global broadcast, questions can originate from any session. Split the
single-session QuestionModal into a thin wrapper that iterates all entries in
the questionModals store and a QuestionCard that owns one question draft. Every
pending question is now visible and answerable regardless of which session tab
is active.
